Fix narrowing conversion in `HashMethodPackedString`
`StringViewHash::operator()` returns `size_t`, but `WeakHash32::data` is
a container of `UInt32`. The implicit conversion was caught by
`-Wshorten-64-to-32` in the `Fast test`, `Build (arm_tidy)`, and
`Build (llvm_coverage_build)` jobs. Add an explicit cast to `UInt32`.
The lower 32 bits are equivalent since `CRC32Hash` fills both halves
with the same value.

Report: https://github.com/ClickHouse/ClickHouse/pull/93271

+++ src/Common/ColumnsHashing/HashMethod.h
@@ -283,7 +283,7 @@ struct HashMethodPackedString : public columns_hashing_impl::HashMethodBase<
             if (str.empty())
                 data[i] = 0;
             else if (str.size() <= std::numeric_limits<UInt32>::max())
-                data[i] = StringViewHash()(str);
+                data[i] = static_cast<UInt32>(StringViewHash()(str));
             else
                 data[i] = static_cast<UInt32>(str.size());
         }
